refactor: Refactor UserInputHandler to DefaultUserInputHandler implementing UserInputHandler interface
This commit renames the UserInputHandler class to DefaultUserInputHandler to better reflect its purpose as the default implementation of the UserInputHandler interface. The class has been updated to implement the UserInputHandler interface and provides the necessary implementation for the handleUserInput() method. This change prepares the code for future use of different implementations of the UserInputHandler interface if needed.

Refs #7