refactor: Refactor AddStudentCommand to decouple InputProvider
Modified the AddStudentCommand class to decouple it from the SystemInputProvider by introducing an InputProvider interface.
Replaced the direct instantiation of SystemInputProvider with the InputProvider parameter in the constructor.
Updated the execute() method to use the provided InputProvider for getting user input.
This modification allows for flexibility in providing different implementations of InputProvider and promotes better code extensibility and testability.

Refs #7