feat: Add student ID and modify student addition logic
This commit introduces several changes to the codebase. The AddStudentsCommand class has been modified to simplify the student addition process by directly passing the extracted credentials to the addStudent method of the StudentRepository.

In the Student class, a new field id has been added to uniquely identify each student. The constructor has been updated to include this new field, and the isValid method has been kept the same to ensure the validity of the credentials.

The most significant changes are in the StudentRepository class. The students collection has been changed to a LinkedList<Student> to preserve the order of student insertion. The addStudent method now accepts Credentials as a parameter instead of a Student object. This modification allows for a more straightforward student addition process. Additionally, the logic for checking if an email is already taken has been moved before creating a new student. Finally, an auto-incrementing id is assigned to each student based on the size of the students list, ensuring a unique identifier for each student.

These changes improve the efficiency and maintainability of the codebase by separating the responsibilities of the AddStudentsCommand and StudentRepository classes and providing a clearer and more concise student addition process.

Refs #10