feat: Add ListCommand to display a list of students
This commit adds a new file, ListCommand.java, to the tracker package. The ListCommand class extends the Command class and is responsible for displaying a list of students.

The ListCommand constructor takes an OutputProvider and a StudentRepository as parameters. The OutputProvider is used to print the list of students, and the StudentRepository is used to retrieve the list of students.

The execute method overrides the execute method of the parent class. It retrieves all the students from the StudentRepository and prints each student's ID using the OutputProvider.

This new command allows users of the learning progress tracker application to view a list of all the students.

Refs #10